% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioshiny.R
\docType{package}
\name{bioshiny}
\alias{bioshiny}
\alias{bioshiny-package}
\title{Can be used to create, downlad and start various Shiny applications.}
\description{
Can be used to create, downlad and start various Shiny applications.
}
\seealso{
Useful links:

\url{https://github.com/ngsjs/bioshiny/src/bioshiny}

Report bugs at \url{https://github.com/ngsjs/bioshiny/issues}
}
\author{
Li Jianfeng \url{lee_jianfeng@sjtu.edu.cn}
}
